requirements.txt
streamlit>=1.36
opencv-python-headless>=4.9
numpy>=1.24
pillow>=10.0
requests>=2.31
matplotlib>=3.8
